Provides TypeScript support for SC theme, images
