--- global.d.ts
@@ -1,9 +1,0 @@
-interface SvgrComponent
-  extends React.StatelessComponent<React.SVGAttributes<SVGElement>> {}
-
-declare module '*.svg' {
-  const url: string
-  export const ReactComponent: SvgrComponent
-
-  export default url
-}
