+++ src/components/layout.tsx
@@ -3,11 +3,11 @@ import { StaticQuery, graphql } from 'gatsby'
 import AtomicLayout from 'atomic-layout'
 import { ThemeProvider, createGlobalStyle } from 'styled-components'
 
+import { useThemePreference } from '../hooks/useThemePreference'
 import PageHeader from './PageHeader'
 import PageFooter from './PageFooter'
 import './layout.css'
 import './custom.css'
-import { useThemePreference } from '../hooks/useThemePreference'
 import { ThemeContext } from './ThemeContext'
 import lightTheme from '../themes/light'
 import darkTheme from '../themes/dark'
