--- styled-components.d.ts
@@ -1,7 +1,0 @@
-import theme from './src/theme'
-
-type Theme = typeof theme
-
-declare module 'styled-components' {
-  export interface DefaultTheme extends Theme {}
-}
